feat: move MediaEmbeds logic to shared module (#4586)
